refactor(cli): move the engine-wiring tests out of agent_tests.rs
CI's check-file-size guard failed: this PR's three new precedence tests took
agent_tests.rs from 1881 to 2008 lines, past its grandfathered ceiling. The
guard offers two ways out and is explicit about which applies -- raise the
baseline only when the growth is irreducible (a module declaration in an
already-oversized lib.rs/main.rs), otherwise put the new code in its own
module. 127 lines of new tests are not irreducible, so they move.

They do not move alone. The three new tests are one subject with the routing
they extend (pipeline_worker_model_is_inert_without_the_fix_but_routes_with_it)
and the fallbacks they share cfg_with_engine with (the worker_model_* group),
so agent_tests/engine_wiring.rs takes all of it: what resolve_engine_wiring
picks, and who wins when several things pin a model. That is the same
#[path] pattern agent_tests/usage_completeness.rs already uses.

agent_tests.rs 2008 -> 1710, back under its 1881 ceiling with room to spare.
The moved code is byte-identical; all eight tests run and pass under
agent::tests::engine_wiring::. The baseline is left alone -- the guard only
enforces the ceiling direction, and a shrinking file needs no baseline edit.

Also fixes two broken rustdoc links that are currently failing `cargo doc
-D warnings` on main. Both came in with #768, which merged while its own CI
run was still going: `crate::store::schema::migrate` names a private module,
and public documentation for `recall_tier` cannot link to the private
`crate::candidates::NodeMeta`. Neither is reachable as a link, so both become
plain prose. Unrelated to this PR's subject, but this branch carries main and
therefore carries the breakage.

All six guards, cargo fmt --check, cargo clippy -D warnings, RUSTDOCFLAGS="-D
warnings" cargo doc, and the full workspace suite pass locally.
